Move helper classes for InfluxDB publisher to ocs library
